feat: teacher can delete chapter and section
